Add a properties dock for the Tiled Collision Editor
You can now edit properties on the object group and objects in the
Collision Editor.

A couple of issues:
- It would be nice to be able to select objects even if the Tileset is
external but the MapView is disabled in this case.
- If I try to keep the MapView enabled then the toolbar actions have to
be disabled, except for the ObjectSelection tool. I couldn't get this to
work -- but I'm new to Qt.
- The ObjectSelection tool is no longer just about selection but allows
you to position, scale, and rotate the object that is currently selected
which further complicates this.
- It would also be nice to be able to go from selecting an Object to the
Object Group easily. Right now this is only achieved by selecting a new
tile (in which case the ObjectGroup for that tile is selected in the
PropertyBrowser until you select an object.